Move around files, add better framebuffer size type
